projects/django_projects/image_vault/requirements.txt
Django>=5.0
gunicorn==23.0.0
uvicorn==0.35.0
psycopg2-binary
celery==5.5.3
django-celery-beat==2.8.1
djangorestframework==3.16.0
django-debug-toolbar==5.2.0
Pillow==11.1.0